### 2018-10-08 #31 @whedon
[PRE REVIEW]: nbgrader: A Tool for Creating and Grading Assignments in the Jupyter Notebook
**Submitting author:** @jhamrick (<a href="http://orcid.org/0000-0002-3860-0429">Jessica B. Hamrick</a>)
**Repository:** <a href="https://github.com/jupyter/nbgrader" target ="_blank">https://github.com/jupyter/nbgrader</a>
**Version:** v0.5.4
