### 2018-10-08 #31 @dsblank
(I am listed as an author; is that still ok?)
