### 2018-10-08 #32 @labarba
👋 @dsblank — Thanks for agreeing to review! Here is where the action happens. You have a review checklist above: feel free to open issues, as needed, in the submission repository, mentioning them here for a link-back. Our [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines) explain what to do, but you can also ask questions here and I'll be glad to help.
