### 2018-10-08 #32 @dsblank
Unfortunately I cannot review this paper as I am an author. But please add me to the list of willing reviewers!
