### 2018-10-08 #31 @jhamrick
Sorry for the confusion, I suggested @dsblank before we had actually finalized the author list so yeah I would definitely think that is now a COI.

Thanks for volunteering @MattForshaw !
