### 2018-10-08 #31 @MattForshaw
> Thanks @MattForshaw! You're familiar with nbgrader already? Can you briefly tell me your background to review this?

No existing familiarity with nbgrader, but plenty of experience using notebooks as a data science educator, so I hope that would be something useful I could bring to a review.
