### 2018-10-08 f25389b
Update FAQ with instructions to fix timeout error.
