### 2018-10-08 fe3faab
FAQ update for instructions to fix timeout errors
Fixed formatting with this commit.
