### 2018-10-09 #31 @labarba
You're good to go! The review will occur at https://github.com/openjournals/jose-reviews/issues/32 and I will now close this issue. Thanks!
