### 2018-10-18 21ad25e
fix: return info of reper function is wrong in api.py
