### 2018-10-24 9c9079f
octave kernel validate now works, needs little cleanup
