### 2018-10-25 f939416
add lms_user_id to the Student object and dict representation
