+++ nbgrader/api.py
@@ -369,6 +369,10 @@ class Student(Base):
     #: of each assignment.
     max_score = None
 
+    #: The lms user ID, this is mainly for identifying students in your LMS system
+    #: and was added so teachers and TA's can easily send grades to an LMS such as CanvasLMS.
+    lms_user_id = None
+
     def to_dict(self):
         """Convert the student object to a JSON-friendly dictionary
         representation.
@@ -381,6 +385,7 @@ class Student(Base):
             "email": self.email,
             "score": self.score,
             "max_score": self.max_score
+            "lms_user_id": self.lms_user_id
         }
 
     def __repr__(self):
