### 2018-10-25 0d58149
adds lms_user_id to the nbgrader student add command
