### 2018-10-29 81e2da9
adds alembic update method for the db to update+downgrade
