+++ nbgrader/alembic/versions/282d30a0218e_adds_lms_user_id_column_to_student_table.py
@@ -1,0 +1,24 @@
+"""adds lms_user_id column to student table
+
+Revision ID: 282d30a0218e
+Revises: 724cde206c17
+Create Date: 2018-10-25 20:14:56.887686
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '282d30a0218e'
+down_revision = '724cde206c17'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    op.add_column('student', sa.Column('lms_user_id', sa.String(128), unique=True, nullable=True))
+
+
+def downgrade():
+    op.drop_column('student', 'lms_user_id')
