### 2018-10-30 dffad00
adds octave support and refactors to less code, also cutoff for errors
