### 2018-10-30 5ff843a
add self to function call
