### 2018-10-31 a5177fb
splitting excepting errors into two separate instead of excepting two in one
