+++ nbgrader/api.py
@@ -1441,19 +1441,16 @@ class Gradebook(object):
                     student=student_id,
                     group_name=group_name
                 ))
-            except (utils.JupyterhubEnvironmentError, utils.JupyterhubApiError) as e:
-                #self.log.error("Error caught: " + e) # self.log not working. Gradebook has no attribute log
-                if self.course_id: # we assume user might be using Jupyterhub but something is not working
-                    # log.error 
-                    # Note: check if log ever appears.
+            except utils.JupyterhubEnvironmentError as e:
+                print("Not running on Jupyterhub, not adding {student} user to Jupyterhub group {group_name}".format(student=student_id, group_name=group_name))
+            except utils.JupyterhubApiError as e:
+                if self.course_id: # We assume user might be using Jupyterhub but something is not working
                     err_msg = "Student {student} NOT added to Jupyterhub group {group_name}: ".format(
                         student=student_id,
                         group_name=group_name
                     )
-                    print(err_msg + str(e))
-                    #self.log.error(err_msg)
+                print(err_msg + str(e))
                 print("Make sure you set a valid api_token in your config file before starting the service")
-                #self.log.error("Make sure you set a valid api_token in your config file before starting the service")
         except (IntegrityError, FlushError) as e:
             self.db.rollback()
             raise InvalidEntry(*e.args)
