### 2018-10-31 11fdcf1
also seperating in remove_student in api.py
