+++ nbgrader/api.py
@@ -1536,13 +1536,14 @@ class Gradebook(object):
                                      api_path="/groups/{name}/users".format(name=group_name),
                                      post_data = {"users":[student.id]}
                 )
-                # log.info
-                print("Student {student} removed or was not in the Jupyterhub group {group_name}".format(student=name, group_name=group_name))
-            except (utils.JupyterhubEnvironmentError, utils.JupyterhubApiError) as e:
-                if self.course_id: # We assume user might be using Jupyterhub but something is not working
-                    # log.error
-                    print("Student {student} NOT removed from the Jupyterhub group {group_name}: ".format(student=name, group_name=group_name) + str(e))
+                print("Student {student} was removed or was already not in the Jupyterhub group {group_name}".format(student=name, group_name=group_name))
+            except utils.JupyterhubEnvironmentError as e:
+                print("Not running on Jupyterhub so {student} was NOT removed from the Jupyterhub group {group_name}: ".format(student=name, group_name=group_name), str(e))
                 #self.log.error("Error caught in remove_student(): " + e)
+            except utils.JupyterhubApiError as e:
+                if self.course_id:
+                    print("Student {student} was NOT removed from the Jupyterhub group {group_name}: ".format(student=name, group_name=group_name), str(e))
+                    print("Make sure you start your service with a valid 'api_token' in your Jupyterhub config")
         except (IntegrityError, FlushError) as e:
             self.db.rollback()
             raise InvalidEntry(*e.args)
