### 2018-10-31 10086d9
adds a check if the group first exists in the add_student, otherwise the post request to a add a group with the group already being there would result in a bad request and a JupyterhubApiError being thrown
