### 2018-10-31 d5d59f5
fixing last commit, add colon and print that group was created
