### 2018-10-31 5fbccb9
removing the log again because Gradebook object has no attribute log
