+++ nbgrader/api.py
@@ -1543,11 +1543,11 @@ class Gradebook(object):
                 )
                 print("Student {student} was removed or was already not in the Jupyterhub group {group_name}".format(student=name, group_name=group_name))
             except utils.JupyterhubEnvironmentError as e:
-                print("Not running on Jupyterhub so {student} was NOT removed from the Jupyterhub group {group_name}: ".format(student=name, group_name=group_name), str(e))
+                print("Not running on Jupyterhub so {student} was NOT removed from the Jupyterhub group {group_name}:".format(student=name, group_name=group_name), str(e))
                 #self.log.error("Error caught in remove_student(): " + e)
             except utils.JupyterhubApiError as e:
                 if self.course_id:
-                    print("Student {student} was NOT removed from the Jupyterhub group {group_name}: ".format(student=name, group_name=group_name), str(e))
+                    print("Student {student} was NOT removed from the Jupyterhub group {group_name}:".format(student=name, group_name=group_name), str(e))
                     print("Make sure you start your service with a valid 'api_token' in your Jupyterhub config")
         except (IntegrityError, FlushError) as e:
             self.db.rollback()
