### 2018-10-31 2bf97b9
fixing log msg I did that accidentally had commas
