### 2018-10-31 f06028d
taking out import urllib.parse from utils.py since it was not being used
