### 2018-10-31 136c426
Refactoring code into new add_student_to_group function and also calling it from update_or_create student
