### 2018-10-31 5500c01
fixes to last commit
