### 2018-11-12 #32 @labarba
Hi @MattForshaw, @jedbrown — 👋
Wondering here what your progress might be with this review for JOSE. Thanks!
