### 2018-11-19 #32 @MattForshaw
Hi @labarba, I've set aside two hours tonight to finish the rest of the checklist for you. Hope this is okay.
