### 2018-11-20 #32 @MattForshaw
Hi @labarba, I am very happy to say nbgrader has passed with flying covers. An impressive submission.
