### 2018-11-22 jupyter/nbgrader#1044 @jedbrown
pyenchant is unmaintained
There doesn't appear to be anyone stepping forward to take over maintenance. This dependency may be an issue going forward. https://github.com/rfk/pyenchant/network

Cc: https://github.com/openjournals/jose-reviews/issues/32
