### 2018-11-22 jupyter/nbgrader#1046 @jedbrown
Public side-effects of running tests
Running the test suite currently tries to create/modify a `sitecustomize.py` in a system path. This file is often not owned by the user and in any case, has public side-effects if other python programs run while the tests run. Instead of these manipulations in `tasks.py`, I think it would be preferable to just have a static file in the nbgrader source and add its directory to PYTHONPATH. Do you foresee any problems with this?

Cc: https://github.com/openjournals/jose-reviews/issues/32
