### 2018-11-22 352a88c
JOSE paper bib updates
Capitalization fixes and add missing DOI.
