### 2018-11-22 #32 @jedbrown
Is there documentation of the branching scheme? The branch `0.5.x` is neither a subset nor superset of `master`. Most branching schemes would merge one way or the other. Also, the tags are "lightweight" as opposed to "annotated" so `git describe` fails without the `--tags` option. Also, the paper is build from `master` which is not a descendant of `v0.5.4`.

I filed a couple issues relating to testing. They are not mandatory to address at this time, but I'd like authors to have a chance to respond.

Some references need capitalization help (see PR).
