### 2018-12-06 2edb4e0
update python versions in travis
