+++ .travis.yml
@@ -5,6 +5,8 @@ python:
   - 2.7
   - 3.5
   - 3.6
+  - 3.7
+  - nightly
 env:
   global:
     - MOZ_HEADLESS=1
