### 2018-12-06 fafbbdf
remove python versions during testing - temporary
