### 2018-12-06 2d3bb82
use recent travis structure
