+++ .travis.yml
@@ -1,6 +1,4 @@
 language: python
-sudo: false
-dist: trusty
 python:
   - 2.7
   - 3.6
