### 2018-12-06 #32 @jhamrick
Hi Jed and Matt, thanks for the review! Just wanted to send a quick note to
say I am a bit overwhelmed with other stuff at the moment, but will try to
address your PRs soon. (Thanks @willinc and @dsblank for handling some of
this already)

The branching scheme is based on the one IPython and I believe some of the
other Jupyter projects use, which is to create release branches and
backport PRs onto them from master. For example, after 0.5.0 was released,
we start developing stuff on master for the 0.6.0 release including major
backward incompatible changes. But occasionally there are bug fixes that
warrant another release sooner. Thus we backport the smaller bugfix changes
to 0.5.x to create intermediate releases 0.5.1, 0.5.2, etc.
