### 2018-12-06 jupyter/nbgrader#1046 @willingc
Yeah, it would be good to remove the usage of `sitecustomize.py`. Probably makes sense to refactor the test runs as well. It would be nice to remove the dependency on invoke too.
