+++ .travis.yml
@@ -1,7 +1,10 @@
 language: python
 python:
   - 2.7
+  - 3.4
+  - 3.5
   - 3.6
+  - nightly
 env:
   global:
     - MOZ_HEADLESS=1
