### 2018-12-06 #32 @jedbrown
Okay, fair enough regarding branches [^1]. I'm more familiar with JOSS, but presume JOSE is similar in that upon acceptance, you would tag a release of the post-review version including the paper sources and archive that to Zenodo. If that's the case and you wish to review the 0.5.x branch (as listed above), I think it would be appropriate to move the paper into the 0.5.x branch so that it will be part of the archive.

IPython uses annotated tags and I would recommend it for new nbgrader tags.

[^1]: I usually branch from the maintenance branch if I'm doing a bug fix, and always merge into 'master', thus avoiding need for a cherry-picked backport. A cosmetic `git merge --strategy=ours 0.5.x` would tie up the graph and make `git describe` more sensible. Anyway, this is wildly tangential.
