### 2018-12-14 57aa50f
fixed PEP8 problems
