+++ nbgrader/tests/api/test_models.py
@@ -209,9 +209,9 @@ def test_create_submitted_assignment(db):
     assert sa.max_written_score == 0
     assert not sa.needs_manual_grade
 
-    assert sa.duedate == None
-    assert sa.timestamp == None
-    assert sa.extension == None
+    assert sa.duedate is None
+    assert sa.timestamp is None
+    assert sa.extension is None
     assert sa.total_seconds_late == 0
 
     d = sa.to_dict()
@@ -242,7 +242,7 @@ def test_submission_timestamp_ontime(db):
 
     assert sa.duedate == duedate
     assert sa.timestamp == timestamp
-    assert sa.extension == None
+    assert sa.extension is None
     assert sa.total_seconds_late == 0
 
 
@@ -258,7 +258,7 @@ def test_submission_timestamp_late(db):
 
     assert sa.duedate == duedate
     assert sa.timestamp == timestamp
-    assert sa.extension == None
+    assert sa.extension is None
     assert sa.total_seconds_late == 172800
 
 
@@ -348,7 +348,7 @@ def test_create_code_grade(db):
     assert g.cell == gc
     assert g.notebook == sn
     assert g.auto_score == 5
-    assert g.manual_score == None
+    assert g.manual_score is None
     assert g.assignment == sa
     assert g.student == s
     assert g.max_score == 10
@@ -407,8 +407,8 @@ def test_create_written_grade(db):
     assert g.id
     assert g.cell == gc
     assert g.notebook == sn
-    assert g.auto_score == None
-    assert g.manual_score == None
+    assert g.auto_score is None
+    assert g.manual_score is None
     assert g.assignment == sa
     assert g.student == s
     assert g.max_score == 10
@@ -473,6 +473,7 @@ def test_create_comment(db):
 
     assert repr(c) == "Comment<foo/blah/foo for 12345>"
 
+
 def test_query_needs_manual_grade_ungraded(submissions):
     db = submissions[0]
 
@@ -516,6 +517,7 @@ def test_query_needs_manual_grade_ungraded(submissions):
         .order_by(api.SubmittedAssignment.id)\
         .all()
     assert a == b
+
 
 def test_query_needs_manual_grade_autograded(submissions):
     db, grades, _ = submissions
@@ -589,6 +591,7 @@ def test_query_needs_manual_grade_autograded(submissions):
         .join(api.SubmittedNotebook, api.Grade)\
         .filter(api.SubmittedNotebook.needs_manual_grade)\
         .all()
+
 
 def test_query_needs_manual_grade_manualgraded(submissions):
     db, grades, _ = submissions
@@ -664,6 +667,7 @@ def test_query_needs_manual_grade_manualgraded(submissions):
         .filter(api.SubmittedNotebook.needs_manual_grade)\
         .all()
 
+
 def test_query_max_score(submissions):
     db = submissions[0]
 
@@ -675,6 +679,7 @@ def test_query_max_score(submissions):
     assert [15, 15] == sorted([x[1] for x in db.query(api.SubmittedAssignment.id, api.SubmittedAssignment.max_score).all()])
     assert [15, 15] == sorted([x[1] for x in db.query(api.Student.id, api.Student.max_score).all()])
 
+
 def test_query_score_ungraded(submissions):
     db = submissions[0]
 
@@ -683,10 +688,12 @@ def test_query_score_ungraded(submissions):
     assert [x[1] for x in db.query(api.SubmittedAssignment.id, api.SubmittedAssignment.score).all()] == [0.0, 0.0]
     assert [x[1] for x in db.query(api.Student.id, api.Student.score).all()] == [0.0, 0.0]
 
+
[…]
