+++ nbgrader/utils.py
@@ -173,19 +173,22 @@ def check_directory(path, read=False, write=False, execute=False):
     else:
         return False
 
+
 def get_osusername():
     """Get the username of the current process."""
     if pwd is None:
         raise OSError("get_username cannot be called on Windows")
     return pwd.getpwuid(os.getuid())[0]
 
+
 def get_username():
     """ Get the username, use os user name but override if username is jovyan ."""
     osname = get_osusername()
-    if osname=='jovyan':
-        return os.environ.get('JUPYTERHUB_USER','jovyan')
+    if osname == 'jovyan':
+        return os.environ.get('JUPYTERHUB_USER', 'jovyan')
     else:
         return osname
+
 
 def find_owner(path):
     """Get the username of the owner of path."""

[…]
