### 2018-12-14 9597563
fixed problems introduced by sticking with pep8
