### 2018-12-15 f2984e4
Backport PR #973: Create a paper on nbgrader
I am very excited to announce that we planning to submit a paper on nbgrader to the [Journal of Open Source Education](http://jose.theoj.org/about#author_guidelines)!

This PR contains the first draft of the paper which will be submitted. For this submission we are following the [Project Jupyter guidelines for JOSS papers](https://github.com/jupyter/governance/blob/master/papers.md#process-for-journal-of-open-source-software-joss). I will be the coordinator and organize the writing and submission of the paper. I have included authors for the first d
[…]
