+++ paper/paper.bib
@@ -1,0 +1,146 @@
+@manual{JupyterHub,
+author = {{Jupyter Development Team}},
+title = {{JupyterHub}},
+year = {2018}
+}
+
+@inproceedings{kluyver2016jupyter,
+  title={Jupyter Notebooks-a publishing format for reproducible computational workflows.},
+  author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
+  booktitle={ELPUB},
+  pages={87--90},
+  year={2016}
+}
+
+@article{Cortina2007,
+author = {Cortina, Thomas J.},
+doi = {10.1145/1227504.1227387},
+isbn = {1595933611},
+issn = {00978418},
+journal = {ACM SIGCSE Bulletin},
+number = {1},
+pages = {218},
+title = {{An introduction to computer science for non-majors using principles of computation}},
+volume = {39},
+year = {2007}
+}
+
+@article{Forte2005,
+author = {Forte, Andrea and Guzdial, Mark},
+journal = {IEEE Transactions on Education},
+number = {2},
+pages = {248--253},
+title = {{Motivation and non-majors in computer science: Identifying discrete audiences for introductory courses}},
+volume = {48},
+year = {2005}
+}
+
+@article{Guzdial2005,
+author = {Guzdial, Mark and Forte, Andrea},
+doi = {10.1145/1047124.1047468},
+isbn = {1581139977},
+issn = {00978418},
+journal = {ACM SIGCSE Bulletin},
+number = {1},
+pages = {361},
+title = {{Design process for a non-majors computing course}},
+volume = {37},
+year = {2005}
+}
+
+@manual{mathematica,
+  Author={{Wolfram Research, Inc.}},
+  Title={Mathematica},
+  Address={Champaign, IL},
+  year = {2018}
+}
+
+@manual{maple,
+  title = {Maple},
+  author = {{Maplesoft, a division of Waterloo Maple Inc.}},
+  Address = {Waterloo, Ontario}
+}
+
+@manual{matlab,
+  Author={{The MathWorks, Inc.}},
+  Title={{MATLAB}},
+  Address={Natick, MA},
+  year = {2018}
+}
+
+@manual{sage,
+  Key          = {Sage},
+  Author       = {W.\thinspace{}A. Stein and others},
+  Organization = {The Sage Development Team},
+  Title        = {{S}age {M}athematics {S}oftware},
+  Year         = {2018},
+}
+
+@Article{PerezGranger2007,
+  Author    = {P\'erez, Fernando and Granger, Brian E.},
+  Title     = {{IP}ython: a System for Interactive Scientific Computing},
+  Journal   = {Computing in Science and Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {21--29},
+  month     = may,
+  year      = 2007,
+  url       = "https://ipython.org",
+  ISSN      = "1521-9615",
+  doi       = {10.1109/MCSE.2007.53},
+  publisher = {IEEE Computer Society},
+}
+
+@misc{Hamrick2016,
+  author       = {Hamrick, Jessica B. and Jupyter Development Team},
+  title        = {2016 Jupyter Education Survey},
+  month        = may,
+  year         = 2016,
+  doi          = {10.5281/zenodo.51701},
+  url          = {https://github.com/jupyter/datasets/tree/master/surveys/2016-05-education-survey}
+}
+
+@misc{Castano_Jupyter_Map,
+  author       = {Castaño, Ezequiel Leonardo},
[…]
