### 2018-12-15 88c356c
Backport PR #1049: Update test builds on Travis
- remove cov-core per @jedbrown #1045
- add more recent python versions
- update to newer travis build parameters
