+++ .travis.yml
@@ -1,10 +1,10 @@
 language: python
-sudo: false
-dist: trusty
 python:
   - 2.7
+  - 3.4
   - 3.5
   - 3.6
+  - nightly
 env:
   matrix:
     - GROUP=docs
