### 2018-12-15 b24f067
Backport PR #1045: Dev requirements and spelling tests
I don't know whether this PR should go to 'master' or '0.5.x'. Should Travis tests be made to regard spelling errors as errors? (It appears that you have to check the logs to see if any errors are identified.)
