+++ nbgrader/docs/source/spelling_wordlist.txt
@@ -205,6 +205,7 @@ student’s
 subclasses
 subcommand
 subcommands
+subdirectory
 suchow
 suprocesses
 svurens
@@ -230,6 +231,7 @@ url
 username
 usernames
 utcnow
+validator
 ve
 vinaykola
 webapp
