### 2018-12-15 e7d4291
Backport PR #1047: JOSE paper bib updates
Capitalization fixes and add missing DOI.

Does the JupyterHub reference have a URL or DOI?

Cc: https://github.com/openjournals/jose-reviews/issues/32
