### 2018-12-15 1f712c5
Backport PR #1016: Fix anaconda link
Fix of #1015

I did a quick look and found only anaconda links are broken, Let me know if there are any additional links which are also broken
