+++ nbgrader/_version.py
@@ -1,2 +1,2 @@
-version_info = (0, 5, 5, 'dev')
+version_info = (0, 5, 5)
 __version__ = '.'.join(map(str, version_info))
