+++ nbgrader/docs/source/spelling_wordlist.txt
@@ -114,6 +114,7 @@ I’m
 I’ve
 jcsutherland
 jdfreder
+jedbrown
 jhamrick
 jilljenn
 jklymak
@@ -209,6 +210,7 @@ subdirectory
 suchow
 suprocesses
 svurens
+swarnava
 sys
 systemwide
 Szepi
