### 2018-12-15 #32 @jhamrick
I've created a new release (0.5.5) with the nbgrader paper on the 0.5.x branch, so that there is a release that corresponds with the paper being reviewd. See: https://github.com/jupyter/nbgrader/pull/1050

I'll keep that branch open until the review here is complete, in case there's any last changes that need to go into it. Thanks for all the suggestions!
