### 2018-12-15 jupyter/nbgrader#1046 @jhamrick
Agreed, all that stuff is there back from when we were running a lot of subprocesses in the tests which weren't by default getting test coverage. But I think that shouldn't be an issue anymore, so we can remove everything having to do with sitecustomize.py.

I'm fine removing the dependency on invoke, too---it's pretty lightweight at this point and we could do everything just with standard argument parsing I think.
