### 2018-12-15 jupyter/nbgrader#1044 @jhamrick
Agreed, I think we can just remove the spell checking from the docs and then also remove pyenchant.
