### 2018-12-15 3bbd482
Remove spellcheck and enchant dependency
