+++ nbgrader/docs/source/conf.py
@@ -34,13 +34,6 @@ extensions = [
     'sphinx.ext.napoleon'
 ]
 
-try:
-    import sphinxcontrib.spelling
-except ImportError:
-    pass
-else:
-    extensions.append("sphinxcontrib.spelling")
-
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
