+++ nbgrader/docs/source/contributor_guide/installation_developer.rst
@@ -15,11 +15,6 @@ nbgrader installs and builds with one command::
 
     pip install -r dev-requirements.txt -e .
 
-Currently, building docs is not supported on Windows because some of the dependencies (enchant)
-are not easily installable. Instead of the above command, run the following on windows::
-
-    pip install -r dev-requirements-windows.txt -e .
-
 
 Installing notebook extensions
 ------------------------------
