### 2018-12-15 59cafbe
Remove customizations to sitecustomize.py
