### 2018-12-15 d505258
Remove dependency on invoke
