+++ tasks.py
@@ -1,30 +1,24 @@
+#!/usr/bin/env python
+
 import os
-from invoke import task, collection
 import sys
+import subprocess as sp
+import argparse
+
+def echo(msg):
+    print("\033[1;37m{0}\033[0m".format(msg))
+
+
+def run(cmd):
+    echo(cmd)
+    return sp.call(cmd, shell=True)
 
 try:
     from nbformat import read
 except ImportError:
     echo("Warning: nbformat could not be imported, some tasks may not work")
 
-if sys.platform == 'win32':
-    WINDOWS = True
-else:
-    WINDOWS = False
-
-
-def run(ctx, *args, **kwargs):
-    if 'pty' not in kwargs:
-        kwargs['pty'] = True
-    if WINDOWS:
-        kwargs['pty'] = False
-    if 'echo' not in kwargs:
-        kwargs['echo'] = True
-    return ctx.run(*args, **kwargs)
-
-
-def echo(msg):
-    print("\033[1;37m{0}\033[0m".format(msg))
+WINDOWS = sys.platform == 'win32'
 
 
 def _check_if_directory_in_path(pth, target):
@@ -34,22 +28,23 @@ def _check_if_directory_in_path(pth, target):
             return True
     return False
 
-@task
-def docs(ctx):
+
+def docs(args):
+    del args  # unused
     if not WINDOWS:
-        run(ctx, 'py.test --nbval-lax --current-env nbgrader/docs/source/user_guide/*.ipynb')
-    run(ctx, 'python nbgrader/docs/source/build_docs.py')
-    run(ctx, 'python nbgrader/docs/source/clear_docs.py')
-    run(ctx, 'make -C nbgrader/docs html')
-    run(ctx, 'make -C nbgrader/docs linkcheck')
+        run('py.test --nbval-lax --current-env nbgrader/docs/source/user_guide/*.ipynb')
+    run('python nbgrader/docs/source/build_docs.py')
+    run('python nbgrader/docs/source/clear_docs.py')
+    run('make -C nbgrader/docs html')
+    run('make -C nbgrader/docs linkcheck')
 
 
-@task
-def cleandocs(ctx):
-    run(ctx, 'python nbgrader/docs/source/clear_docs.py')
+def cleandocs(args):
+    del args  # unused
+    run('python nbgrader/docs/source/clear_docs.py')
 
 
-def _run_tests(ctx, mark=None, skip=None, junitxml=None):
+def _run_tests(mark, skip, junitxml):
     cmd = []
     cmd.append('py.test')
     if not WINDOWS:
@@ -69,72 +64,90 @@ def _run_tests(ctx, mark=None, skip=None, junitxml=None):
     if len(marks) > 0:
         cmd.append('-m "{}"'.format(" and ".join(marks)))
 
-    run(ctx, " ".join(cmd))
+    run(" ".join(cmd))
 
     if not WINDOWS:
-        run(ctx, "ls -a .coverage*")
-        run(ctx, "coverage combine || true")
+        run("ls -a .coverage*")
+        run("coverage combine || true")
 
 
-@task
-def tests(ctx, group='all', skip=None, junitxml=None):
-    if group == 'python':
-        _run_tests(ctx, mark="not nbextensions", skip=skip, junitxml=junitxml)
+def tests(args):
+    if args.group == 'python':
+        _run_tests(
+            mark="not nbextensions", skip=args.skip, junitxml=args.junitxml)
 
-    elif group == 'nbextensions':
-        _run_tests(ctx, mark="nbextensions", skip=skip, junitxml=junitxml)
+    elif args.group == 'nbextensions':
[…]
