### 2018-12-15 54f9185
Change 'py.test' to 'pytest'
