### 2018-12-15 2dec533
Don't fail on coverage combine
