### 2018-12-15 8a14462
Don't rely on https://github.com/minrk/travis-wheels, which is now deprecated
