### 2018-12-15 8a0aa09
Replace last few invoke calls in appveyor.yml
