### 2018-12-15 7718ea9
Remove dependency on https://github.com/minrk/travis-wheels, which is now deprecated
