### 2018-12-15 8e0e6b8
Pin ipython to version 6.2.1
