+++ setup.py
@@ -102,7 +102,8 @@ setup_args = dict(
         "six>=1.9",
         "requests",
         "jsonschema",
-        "alembic"
+        "alembic",
+        "ipython==6.2.1"
     ]
 )
 
