### 2018-12-15 a0bce89
Require at most ipython 6.2.1 rather than being strict
