+++ setup.py
@@ -103,7 +103,7 @@ setup_args = dict(
         "requests",
         "jsonschema",
         "alembic",
-        "ipython==6.2.1"
+        "ipython<=6.2.1"
     ]
 )
 
