### 2018-12-15 faed0e4
Also pin version of jupyter-console
