### 2018-12-15 jupyter/nbgrader#1046 @willingc
Nicely done @jhamrick :cake:
