### 2018-12-16 c68951b
Apply suggestions from code review
Added @jhamrick suggestions

Co-Authored-By: danielmaitre <daniel.maitre@durham.ac.uk>
