### 2018-12-16 #32 @labarba
I see only a couple of unchecked item in the review checklist. Are there more change requests pending?
