### 2018-12-16 #32 @jhamrick
I think I have managed to address all the comments at this point---unless
there  is anything else on @jedbrown 's end the only thing remaining is for
me to make the 0.5.5 release and update the version here to be 0.5.5. But I
would just like to hear from him first before I do that in case there's
anything else that needs to go in!

On Sun, Dec 16, 2018, 10:31 PM Lorena A. Barba <notifications@github.com
wrote:

> I see only a couple of unchecked item in the review checklist. Are there
> more change requests pending?
>
