### 2018-12-27 #32 @jedbrown
Looks good to me. Looks like the housekeeping mainly landed in 'master', but it wasn't essential to the paper anyway. My understanding is that you will tag `v0.5.5` on branch '0.5.x' upon acceptance (prior to archival on Zenodo). Regarding authorship, I assume that any listed authors that have not authored commits in the repository have made other significant contributions to the project and that authors of a modest number of commits (such as Auberon Lopez, Lucas Kushner, and Ella Chao) are more appropriate to recognize collectively under Project Jupyter. (I checked the box under this assumption; the core team is certainly among the authors listed for the paper.)

My last nit would be to check capitalization of "notebook" for consistency. (The Jupyter4Edu consensus was that "Jupyter Notebook" refers to the software while "Jupyter notebook" is a notebook.)
