### 2019-01-01 9d3e9dd
groupshared: use c.CourseDir.groupshared option
- Instead of separate assignment/exchange options, which was redundant
  and a bad design in the first place.
- Thanks to @jhamrick for the pointer in review.
