### 2019-01-02 #32 @jhamrick
Yes, I just got back from the holidays and will fix the capitalization
tonight and make the 0.5.5 release tonight!

On Mon, Dec 31, 2018 at 3:25 PM Lorena A. Barba <notifications@github.com>
wrote:

> @jhamrick <https://github.com/jhamrick> 👋 ... could you tweak the
> capitalization of "notebook," like Jed asks? Then we can move to accept
> this submission!
>
