### 2019-01-02 #32 @jhamrick
> Regarding authorship, I assume that any listed authors that have not authored commits in the repository have made other significant contributions to the project and that authors of a modest number of commits (such as Auberon Lopez, Lucas Kushner, and Ella Chao) are more appropriate to recognize collectively under Project Jupyter.

Yes. In determining authorship I reached out to everyone who had made commits to the project and asked who wanted to be included as an author, following the process outlined here: https://github.com/jupyter/governance/blob/master/papers.md#process-for-journal-of-open-source-software-joss

See https://github.com/jupyter/nbgrader/pull/973 for all the discussion on author contributions and the writing of the paper itself.

> My last nit would be to check capitalization of "notebook" for consistency. (The Jupyter4Edu consensus was that "Jupyter Notebook" refers to the software while "Jupyter notebook" is a notebook.)

See: https://github.com/jupyter/nbgrader/pull/1057
