### 2019-01-03 #32 @labarba
OK! Since we're ready to publish, would you do the merging kung fu that you have planned, then make a Zenodo deposit of the whole repo, and post the DOI here?
