### 2019-01-03 7e704a7
Backport PR #1057: Ensure consistency in capitalizing Jupyter Notebook
This makes the capitalization of "Jupyter Notebook" more consistent, as per https://github.com/openjournals/jose-reviews/issues/32#issuecomment-450084190 .

Can one of @willingc @dsblank @lgpage @ellisonbg take a second look just to make sure I didn't make a typo somewhere?

cc @jedbrown
