+++ paper/paper.md
@@ -80,27 +80,27 @@ bibliography: paper.bib
 # Summary
 
 nbgrader is a flexible tool for creating and grading assignments in the Jupyter
-notebook [@kluyver2016jupyter]. nbgrader allows instructors to create a single,
+Notebook [@kluyver2016jupyter]. nbgrader allows instructors to create a single,
 master copy of an assignment, including tests and canonical solutions. From the
 master copy, a student version is generated without the solutions, thus
 obviating the need to maintain two separate versions. nbgrader also
 automatically grades submitted assignments by executing the notebooks and
 storing the results of the tests in a database. After auto-grading, instructors
 can manually grade free responses and provide partial credit using the
-*formgrader* notebook extension. Finally, instructors can use nbgrader to leave
-personalized feedback for each student's submission, including comments as well
-as detailed error information.
+*formgrader* Jupyter Notebook extension. Finally, instructors can use nbgrader
+to leave personalized feedback for each student's submission, including comments
+as well as detailed error information.
 
 nbgrader can also be used with JupyterHub [@JupyterHub], which is a centralized,
 server-based installation that manages user logins and management of Jupyter
-notebooks. When used with JupyterHub, nbgrader provides additional workflow
-functionality, covering the entire grading process. After creating an
+Notebook servers. When used with JupyterHub, nbgrader provides additional
+workflow functionality, covering the entire grading process. After creating an
 assignment, instructors can distribute it to students, who can then fetch a copy
-of the assignment directly through the notebook server interface. Students can
-submit their completed assignment through the same interface, making it
-available for instructors. After students submit their assignments, instructors
-can collect the assignments with a single command and use the auto-grading
-functionality in the normal way.
+of the assignment directly through the Jupyter Notebook server interface.
+Students can submit their completed assignment through the same interface,
+making it available for instructors. After students submit their assignments,
+instructors can collect the assignments with a single command and use the
+auto-grading functionality in the normal way.
 
 Since its conception in September 2014, nbgrader has been used in a number of
 educational contexts, including courses at UC Berkeley, Cal Poly, University of
@@ -136,52 +136,52 @@ for distributing, collecting, and evaluating student work.
 In recent years, the IPython project introduced the *Jupyter Notebook*
 [@kluyver2016jupyter], an interface that is conducive to interactive and
 literate computing, where programmers can interleave prose with code, equations,
-figures, and other media. Jupyter notebooks were originally developed for
+figures, and other media. The Jupyter Notebook was originally developed for
 programming in the Python programming language, but multiple languages are now
 supported using the same infrastructure. The Jupyter Notebook is ideal for
 educators because it allows them to create assignments which include
 instructions along with code or Markdown cells, in which students can provide
 solutions to exercises. Students can, for example, write code both to compute
-and visualize a particular result. Because the notebook is interactive, students
-can iterate on a coding problem without needing to switch back and forth between
-a command line and a text editor, and they can rapidly see results alongside the
-code which produced them.
+and visualize a particular result. Because the Jupyter Notebook is interactive,
+students can iterate on a coding problem without needing to switch back and
+forth between a command line and a text editor, and they can rapidly see results
+alongside the code which produced them.
 
 Instructors in many fields have begun using the Jupyter Notebook as a teaching
-platform. The notebook has appeared in over 100 classes [@Hamrick2016;
+platform. The Jupyter Notebook has appeared in over 100 classes [@Hamrick2016;
 @Castano_Jupyter_Map; @Castano_Jupyter_Map_Dataset] on subjects including
 geology, mathematics, mechanical engineering, data science, chemical
 engineering, and bioinformatics. Software Carpentry, an organization that aims
[…]
