+++ nbgrader/docs/source/changelog.rst
@@ -14,6 +14,7 @@ A summary of changes to nbgrader.
 nbgrader version 0.5.5 is a release for the Journal of Open Source education,
 with the following PRs merged:
 
+- PR #1057: Ensure consistency in capitalizing Jupyter Notebook
 - PR #1049: Update test builds on Travis
 - PR #1047: JOSE paper bib updates
 - PR #1045: Dev requirements and spelling tests
