### 2019-01-06 #32 @jhamrick
Done! Here is the DOI from zenodo: 10.5281/zenodo.2532780
