### 2019-01-06 0828273
Update changelog for 0.5.5 [ci skip]
