+++ nbgrader/_version.py
@@ -1,2 +1,2 @@
-version_info = (0, 5, 5)
+version_info = (0, 5, 6, 'dev')
 __version__ = '.'.join(map(str, version_info))
