### 2019-01-06 #32 @labarba
Congratulations, everyone!
This iconic JOSE paper is now published.

Huge thanks to @MattForshaw and @jedbrown for their review:
https://twitter.com/LorenaABarba/status/1082023133681463296
